api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
cassandra-driver==3.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
